requirements.txt
python-telegram-bot==11.1.0
dialogflow==1.0.0
python-dotenv==0.13.0
PySocks==1.7.1
